% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa.helpers-package.R
\docType{package}
\name{DHARMa.helpers-package}
\alias{DHARMa.helpers}
\alias{DHARMa.helpers-package}
\title{DHARMa.helpers: Helper functions to check models not (yet) directly supported by DHARMa}
\description{
Helper functions to check models not (yet) directly supported by the DHARMa package,
    such as Bayesian models fitted with brms, or exponential random graph models (ERGMs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pakillo/DHARMa.helpers}
  \item Report bugs at \url{https://github.com/Pakillo/DHARMa.helpers/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodríguez-Sánchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
