% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh_check_brms.R
\name{dh_check_brms}
\alias{dh_check_brms}
\title{Check Bayesian models fitted with brms}
\usage{
dh_check_brms(
  model,
  integer = FALSE,
  plot = TRUE,
  nsamples = 1000,
  ntrys = 5,
  ...
)
}
\arguments{
\item{model}{A fitted model \code{\link[brms:brmsfit-class]{brms::brmsfit-class()}}.}

\item{integer}{Logical (TRUE/FALSE), indicating if response is an integer, as in Poisson and binomial models}

\item{plot}{Logical. Plot residual checks? Default is TRUE.}

\item{nsamples}{Integer. Number of samples to draw from the posterior.}

\item{ntrys}{Integer. Number of trys to use for truncated distributions. See \code{\link[brms:posterior_predict.brmsfit]{brms::posterior_predict()}}.}

\item{...}{Further arguments for \code{\link[DHARMa:plotResiduals]{DHARMa::plotResiduals()}}}
}
\value{
An object of type \code{DHARMa}. See \code{\link[DHARMa:createDHARMa]{DHARMa::createDHARMa()}} for more details.
}
\description{
Check Bayesian models fitted with brms
}
\examples{
\dontrun{
# Poisson regression for the number of seizures in epileptic patients
fit1 <- brm(count ~ zAge + zBase * Trt + (1|patient),
           data = epilepsy, family = poisson())
simres <- dh_check_brms(fit1, integer = TRUE) 
plot(simres, form = epilepsy$zAge)
testDispersion(simres)         
}
}
